requirements.txt
python-telegram-bot==20.6
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
requests==2.31.0
jinja2==3.1.3
python-multipart==0.0.9
PyJWT==2.10.1
